Feat: 로컬스토리지 item 삭제 함수 추가 #176
