+++ src/apis/logOut.tsx
@@ -3,6 +3,7 @@ import { useNavigate } from "react-router-dom";
 import useRefreshTokenApi from "./useRefreshToken";
 import apiInstance from "../utils/apiInstance";
 import isApiError from "../utils/isApiError";
+import removeLocalStorageItem from "../utils/removeLocalStorageItem";
 
 // 로그아웃 커스텀훅_박예선_2023.01.30
 // 단순 로그아웃 시 호출: useLogOut();
@@ -43,16 +44,6 @@ const useLogOut = () => {
 };
 
 export default useLogOut;
-
-function removeLocalStorageItem() {
-  localStorage.removeItem("accessToken");
-  localStorage.removeItem("refreshToken");
-  localStorage.removeItem("memberId");
-  localStorage.removeItem("email");
-  localStorage.removeItem("nickname");
-  localStorage.removeItem("memberImage");
-  localStorage.removeItem("authority");
-}
 
 interface LogOutRes {
   resultCode?: number;
